update acceleration analysis

+++ lib/domain/entities/daily_report.dart
@@ -11,7 +11,7 @@ class DailyReport {
   // minute
   int activeTime;
   // cal
-  int calories;
+  double calories;
   DailyGoal goal;
 
   DailyReport({
